Fixed user_showreps setter

+++ database/migrations/2015_12_11_235606_create_log_exercises_table.php
@@ -22,7 +22,7 @@ class CreateLogExercisesTable extends Migration
             $table->integer('logex_reps');
             $table->integer('logex_sets');
             $table->double('logex_failed_volume', 20, 2); // NEW colomn
-            $table->double('logex_failed_sets', 20, 2); // NEW colomn
+            $table->integer('logex_failed_sets'); // NEW colomn
             $table->double('logex_warmup_volume', 20, 2); // NEW colomn
             $table->integer('logex_warmup_reps'); // NEW colomn
             $table->integer('logex_warmup_sets'); // NEW colomn
